Started playing around with Obstacles. They currently are destroyed when colliding with the player. Have to find a way that allows the collision to happen without any rigid stuff going on while still being able to apply velocity to the Obstacles.

+++ Assets/Logic/Mappings/PMappingConfigs.cs
@@ -44,8 +44,8 @@
     public class PObstacleConfig
     {
         public float time;
-        public float lineIndex;
-        public float type;
+        public int lineIndex;
+        public int type;
         public float duration;
         public float width;
     }
